Make OverflowType public
Not technically necessary (C++ allows a public member with private type, mind-bogglingly! See: https://stackoverflow.com/questions/72511032/is-there-any-point-in-having-a-public-member-with-a-private-type), but I don't want to have a public member with private type. This adds unnecessary complexity in this case, we're fine with OverflowType being exposed (it's just a native type). We could even expose StorageType without issue, but there's currently no use for doing so, so we're not going to expose that other type at this point.